+++ CrispyWaffle.sln
@@ -32,6 +32,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrispyWaffle.Utils", "Src\C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrispyWaffle.Eventlog", "Src\CrispyWaffle.Eventlog\CrispyWaffle.Eventlog.csproj", "{0E0F2FA2-88CE-46FA-9D0B-27F184BD4F24}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "i18n", "i18n", "{EC4D48C9-3453-4A42-8D8E-87940991F187}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrispyWaffle.I18n.PtBr", "Src\CrispyWaffle.I18n.PtBr\CrispyWaffle.I18n.PtBr.csproj", "{7ECDAB18-856C-4E34-83DE-75FFDC7C074F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CrispyWaffle.I18n.PtBr.Tests", "Tests\CrispyWaffle.I18n.PtBr.Tests\CrispyWaffle.I18n.PtBr.Tests.csproj", "{5DE2A8BB-8319-4DB5-8DC9-FC78E98FF685}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -78,6 +84,14 @@ Global
 		{0E0F2FA2-88CE-46FA-9D0B-27F184BD4F24}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0E0F2FA2-88CE-46FA-9D0B-27F184BD4F24}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0E0F2FA2-88CE-46FA-9D0B-27F184BD4F24}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7ECDAB18-856C-4E34-83DE-75FFDC7C074F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7ECDAB18-856C-4E34-83DE-75FFDC7C074F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7ECDAB18-856C-4E34-83DE-75FFDC7C074F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7ECDAB18-856C-4E34-83DE-75FFDC7C074F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5DE2A8BB-8319-4DB5-8DC9-FC78E98FF685}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5DE2A8BB-8319-4DB5-8DC9-FC78E98FF685}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DE2A8BB-8319-4DB5-8DC9-FC78E98FF685}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5DE2A8BB-8319-4DB5-8DC9-FC78E98FF685}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -93,6 +107,9 @@ Global
 		{A4C3BBE2-F33F-4752-97C4-206D7C4D6DDD} = {2FD912BD-34A5-4F9E-8192-CD7AEFD572B5}
 		{C86C7297-1E8D-4611-B01A-08FC4FDF1D62} = {2FD912BD-34A5-4F9E-8192-CD7AEFD572B5}
 		{0E0F2FA2-88CE-46FA-9D0B-27F184BD4F24} = {2FD912BD-34A5-4F9E-8192-CD7AEFD572B5}
+		{EC4D48C9-3453-4A42-8D8E-87940991F187} = {2FD912BD-34A5-4F9E-8192-CD7AEFD572B5}
+		{7ECDAB18-856C-4E34-83DE-75FFDC7C074F} = {EC4D48C9-3453-4A42-8D8E-87940991F187}
+		{5DE2A8BB-8319-4DB5-8DC9-FC78E98FF685} = {AA14B952-E519-4983-BF49-E106490AAF1F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {FAE960DB-B996-4FB3-858C-61BA7C7CD86E}
