+++ Src/CrispyWaffle.Eventlog/Log/Adapters/EventLogAdapter.cs
@@ -3,7 +3,6 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using CrispyWaffle.Eventlog.Log.Providers;
 using CrispyWaffle.Log.Adapters;
-using CrispyWaffle.Log.Providers;
 using CrispyWaffle.Serialization;
 using LogLevel = CrispyWaffle.Log.LogLevel;
