+++ Src/CrispyWaffle/Serialization/Adapters/NewtonsoftJsonSerializerAdapter.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Globalization;
+﻿using System.Globalization;
 using System.IO;
 using System.Text;
 using CrispyWaffle.Serialization.NewtonsoftJson;
